Move comment section to locations page
