--- client/src/comments/index.tsx
@@ -1,11 +1,0 @@
-import Comments from './components/Comments';
-
-const CommentSection = () => {
-  return (
-    <div>
-      <Comments />
-    </div>
-  );
-};
-
-export default CommentSection;
