+++ client/src/pages/LocationPage/CommentSection/index.scss
