+++ client/src/pages/LocationPage/CommentSection/user-icon.png
