+++ client/src/pages/LocationPage/index.tsx
@@ -6,7 +6,7 @@ import { RootState } from '../../reducers';
 import { useParams } from 'react-router-dom';
 import './index.scss';
 import { useState, useEffect } from 'react';
-import Comments from '../../comments/components/Comments';
+import CommentSection from './CommentSection';
 
 function LocationPage() {
   const { id } = useParams<{ id: string }>();
@@ -61,8 +61,7 @@ function LocationPage() {
         </div>
 
         <div className="align-text-left">
-          {/* TODO implement comments */}
-          <Comments />
+          <CommentSection />
         </div>
       </Container>
     </div>
